various fixes & enhancements
