+++ mlrun/__init__.py
@@ -14,7 +14,6 @@
 
 __version__ = '0.2.0'
 
-from .run import get_or_create_ctx, new_function, mlrun_op
+from .run import get_or_create_ctx, new_function, mlrun_op, notebook_function
 from .db import get_run_db
 from .model import RunTemplate, NewRun, RunObject
-from .runtimes import make_nuclio_job
