+++ mlrun/runtimes/__init__.py
@@ -17,5 +17,5 @@ from .local import HandlerRuntime, LocalRuntime
 from .function import RemoteRuntime
 from .mpijob import MpiRuntime, MpiJob
 from .dask import DaskCluster
-from .kubejob import KubejobRuntime, make_nuclio_job
+from .kubejob import KubejobRuntime
 from .nuclio_deploy import NuclioDeployRuntime
