+++ mlrun/runtimes/base.py
@@ -33,6 +33,8 @@ from ..execution import MLClientCtx
 from ..artifacts import TableArtifact
 from ..lists import RunList
 from .generators import GridGenerator, ListGenerator
+from ..k8s_utils import k8s_helper
+from ..builder import build_runtime
 
 
 class RunError(Exception):
@@ -56,6 +58,7 @@ class FunctionSpec(ModelObj):
 
 class RunRuntime(ModelObj):
     kind = 'base'
+    _dict_fields = ['kind', 'metadata', 'spec', 'build']
 
     def __init__(self, metadata=None, spec=None, build=None):
         self._metadata = None
@@ -67,6 +70,8 @@ class RunRuntime(ModelObj):
         self.build = build
         self._db_conn = None
         self._secrets = None
+        self._k8s = None
+        self._is_built = False
 
     @property
     def metadata(self) -> BaseMetadata:
@@ -91,6 +96,15 @@ class RunRuntime(ModelObj):
     @build.setter
     def build(self, build):
         self._build = self._verify_dict(build, 'build', ImageBuilder)
+
+    def _get_k8s(self):
+        if not self._k8s:
+            self._k8s = k8s_helper()
+        return self._k8s
+
+    def set_label(self, key, value):
+        self.metadata.labels[key] = str(value)
+        return self
 
     def with_code(self, from_file='', body=None):
         if (not body and not from_file) or (from_file and from_file.endswith('.ipynb')):
@@ -247,6 +261,48 @@ class RunRuntime(ModelObj):
         if self.spec.args:
             args += self.spec.args
         return command, args
+
+    def build_image(self, image, base_image=None, commands: list = None,
+                    secret=None, with_mlrun=True, watch=True):
+        self.build.image = image
+        self.spec.image = ''
+        if commands and isinstance(commands, list):
+            self.build.commands = self.build.commands or []
+            self.build.commands += commands
+        if secret:
+            self.build.secret = secret
+        if base_image:
+            self.build.base_image = base_image
+        ready = self._build_image(watch, with_mlrun)
+        return self
+
+    def _build_image(self, watch=False, with_mlrun=True, execution=None):
+        pod = self.build.build_pod
+        if not self._is_built and pod:
+            k8s = self._get_k8s()
+            status = k8s.get_pod_status(pod)
+            if status == 'succeeded':
+                self.build.build_pod = None
+                self._is_built = True
+                logger.info('build completed successfully')
+                return True
+            if status in ['failed', 'error']:
+                raise RunError(' build {}, watch the build pod logs: {}'.format(status, pod))
+            logger.info('builder status is: {}, wait for it to complete'.format(status))
+            return False
+
+        if not self.build.commands and self.spec.mode == 'pass' and not self.build.source:
+            if not self.spec.image and not self.build.base_image:
+                raise RunError('image or base_image must be specified')
+            self.spec.image = self.spec.image or self.build.base_image
+        if self.spec.image:
+            self._is_built = True
+            return True
+
+        if execution:
+            execution.set_state('build')
+        ready = build_runtime(self, with_mlrun, watch)
+        self._is_built = ready
 
     def _run(self, runspec: RunObject, execution) -> dict:
         pass
